refactor and improve naming
