+++ hudi-client/hudi-client-common/src/main/java/org/apache/hudi/table/action/index/BaseIndexingCatchupTask.java
@@ -19,10 +19,16 @@
 
 package org.apache.hudi.table.action.index;
 
+import org.apache.hudi.avro.model.HoodieCleanMetadata;
+import org.apache.hudi.avro.model.HoodieRestoreMetadata;
+import org.apache.hudi.avro.model.HoodieRollbackMetadata;
 import org.apache.hudi.client.transaction.TransactionManager;
 import org.apache.hudi.common.engine.HoodieEngineContext;
 import org.apache.hudi.common.table.HoodieTableMetaClient;
 import org.apache.hudi.common.table.timeline.HoodieInstant;
+import org.apache.hudi.common.table.timeline.HoodieTimeline;
+import org.apache.hudi.common.table.timeline.TimelineMetadataUtils;
+import org.apache.hudi.common.util.CleanerUtils;
 import org.apache.hudi.common.util.Option;
 import org.apache.hudi.exception.HoodieIndexException;
 import org.apache.hudi.metadata.HoodieTableMetadataWriter;
@@ -30,9 +36,14 @@ import org.apache.hudi.metadata.HoodieTableMetadataWriter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.io.IOException;
 import java.util.List;
 import java.util.Set;
 
+import static org.apache.hudi.common.table.timeline.HoodieInstant.State.COMPLETED;
+import static org.apache.hudi.common.table.timeline.HoodieTimeline.CLEAN_ACTION;
+import static org.apache.hudi.common.table.timeline.HoodieTimeline.RESTORE_ACTION;
+import static org.apache.hudi.common.table.timeline.HoodieTimeline.ROLLBACK_ACTION;
 import static org.apache.hudi.table.action.index.RunIndexActionExecutor.TIMELINE_RELOAD_INTERVAL_MILLIS;
 
 /**
@@ -71,9 +82,75 @@ public abstract class BaseIndexingCatchupTask implements Runnable {
   }
 
   @Override
-  public abstract void run();
+  public void run() {
+    for (HoodieInstant instant : instantsToIndex) {
+      // metadata index already updated for this instant
+      instant = awaitInstantCaughtUp(instant);
+      if (instant == null) {
+        continue;
+      }
+      // if instant completed, ensure that there was metadata commit, else update metadata for this completed instant
+      if (COMPLETED.equals(instant.getState())) {
+        String instantTime = instant.getTimestamp();
+        Option<HoodieInstant> metadataInstant = metadataMetaClient.reloadActiveTimeline()
+            .filterCompletedInstants().filter(i -> i.getTimestamp().equals(instantTime)).firstInstant();
+        if (metadataInstant.isPresent()) {
+          currentCaughtupInstant = instantTime;
+          continue;
+        }
+        try {
+          // we need take a lock here as inflight writer could also try to update the timeline
+          transactionManager.beginTransaction(Option.of(instant), Option.empty());
+          LOG.info("Updating metadata table for instant: " + instant);
+          switch (instant.getAction()) {
+            case HoodieTimeline.COMMIT_ACTION:
+            case HoodieTimeline.DELTA_COMMIT_ACTION:
+            case HoodieTimeline.REPLACE_COMMIT_ACTION:
+              updateIndexForWriteAction(instant);
+              break;
+            case CLEAN_ACTION:
+              HoodieCleanMetadata cleanMetadata = CleanerUtils.getCleanerMetadata(metaClient, instant);
+              metadataWriter.update(cleanMetadata, instant.getTimestamp());
+              break;
+            case RESTORE_ACTION:
+              HoodieRestoreMetadata restoreMetadata = TimelineMetadataUtils.deserializeHoodieRestoreMetadata(
+                  metaClient.getActiveTimeline().getInstantDetails(instant).get());
+              metadataWriter.update(restoreMetadata, instant.getTimestamp());
+              break;
+            case ROLLBACK_ACTION:
+              HoodieRollbackMetadata rollbackMetadata = TimelineMetadataUtils.deserializeHoodieRollbackMetadata(
+                  metaClient.getActiveTimeline().getInstantDetails(instant).get());
+              metadataWriter.update(rollbackMetadata, instant.getTimestamp());
+              break;
+            default:
+              throw new IllegalStateException("Unexpected value: " + instant.getAction());
+          }
+        } catch (IOException e) {
+          throw new HoodieIndexException(String.format("Could not update metadata partition for instant: %s", instant), e);
+        } finally {
+          transactionManager.endTransaction(Option.of(instant));
+        }
+      }
+    }
+  }
 
-  HoodieInstant getHoodieInstant(HoodieInstant instant) {
+  /**
+   * Updates metadata table for the instant. This is only called for actions that do actual writes,
+   * i.e. for commit/deltacommit/compaction/replacecommit and not for clean/restore/rollback actions.
+   *
+   * @param instant HoodieInstant for which to update metadata table
+   * @throws IOException
+   */
+  public abstract void updateIndexForWriteAction(HoodieInstant instant) throws IOException;
+
+  /**
+   * For the given instant, this method checks if it is already caught up or not.
+   * If not, it waits until the instant is completed.
+   *
+   * @param instant HoodieInstant to check
+   * @return null if instant is already caught up, else the instant after it is completed.
+   */
+  HoodieInstant awaitInstantCaughtUp(HoodieInstant instant) {
     if (!metadataCompletedInstants.isEmpty() && metadataCompletedInstants.contains(instant.getTimestamp())) {
       currentCaughtupInstant = instant.getTimestamp();
       return null;
